Break apart dashboard into seperate modules
